Jktoiminnat

+++ OmaJaakaappi.sql
@@ -7,7 +7,7 @@ CREATE TABLE tuote
   tuote_id INT NOT NULL AUTO_INCREMENT,
   tuote_nimi VARCHAR(100) NOT NULL,
   tuote_yksikko VARCHAR(100),
-  tuote_kcal INT,
+  tuote_kcal DOUBLE,
   PRIMARY KEY (tuote_id)
 );
 
@@ -15,7 +15,7 @@ CREATE TABLE jaakaappi
 (
   jaakaappi_id INT NOT NULL AUTO_INCREMENT,
   tuote_pvm DATE,
-  tuote_maara FLOAT,
+  tuote_maara DOUBLE,
   tuote_status VARCHAR(100) NOT NULL,
   tuote_id INT NOT NULL,
   PRIMARY KEY (jaakaappi_id),
@@ -41,7 +41,7 @@ CREATE TABLE rpk
 CREATE TABLE ainekset
 (
   aines_id INT NOT NULL AUTO_INCREMENT,
-  aines_maara FLOAT,
+  aines_maara DOUBLE,
   tuote_id INT NOT NULL,
   resepti_id INT NOT NULL,
   PRIMARY KEY (aines_id),
